% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2R.R
\name{df2R}
\alias{df2R}
\title{Transform Dataframe to Matrix R}
\usage{
df2R(R, r)
}
\arguments{
\item{R}{dataframe of sampling variances of direct estimators.}

\item{r}{number of variables}
}
\value{
Block diagonal matrix R
}
\description{
This function transforms dataframe contains sampling variance to a diagonal matrix R
}
\examples{
NULL
}
